backend/requirement.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pytest